Adds Bottom Navigation Bar
Adds back Bottom Navigation Bar which is the recommended navigation
system as of the latest material design guidelines.

Removes unncessary classes that served as fragment containers as it is
one of the wrost practises. Instead, uses a single class to make the
whole fragments to contain into thereby removing duplicate code while
promoting OOP concepts.

Fixes #145

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>